+++ test/test-attributes.js
@@ -52,7 +52,7 @@ test('Must override class and style attributes', async t => {
 
 test('Must remove an attributes that has "undefined" or "null" value', async t => {
   const actual = `<component src="components/remove-attributes.html" tabindex="-1">My button</component>`;
-  const expected = `<button class="btn btn-primary" data-bs-dismiss="true" data-bs-backdrop="false" tabindex="-1">My button</button>`;
+  const expected = `<button class="btn btn-primary" data-bs-dismiss="true" data-bs-backdrop="false" tabindex="-1">My button</button><div>works also in all nodes</div>`;
 
   const html = await posthtml([plugin({root: './test/templates', tag: 'component'})]).process(actual).then(result => clean(result.html));